event_manager/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile '.\requirements-dev.in'
#
asgiref==3.7.2
    # via
    #   -c .\requirements.txt
    #   django
django==4.2.2
    # via
    #   -c .\requirements.txt
    #   django-debug-toolbar
django-debug-toolbar==4.1.0
    # via -r .\requirements-dev.in
sqlparse==0.4.4
    # via
    #   -c .\requirements.txt
    #   django
    #   django-debug-toolbar
tzdata==2023.3
    # via
    #   -c .\requirements.txt
    #   django
